Add utility to test scrubcam/scrubhub without graphics
In situations where we are missing a screen or a display-related
dependency, we sometimes need to test that scrubcam can connect to
scrubhub.  This utility fills this need for now.  One weakness of this
method is that any changes to the real scrubhub will need to be
manually added to this utility but the need was pressing enough now to
do this this way for the time being.