feat: Implement heartbeat connection signal
ScrubCam sends a "heartbeat" message to ScrubDash every 15 seconds. This
lets ScrubDash know if ScrubCam is still connected or not when images
are not being sent.

The ALERT_CLASSES and COOLDOWN_TIME keys are removed from the config
file. These are moved to the ScrubDash config file since they act as a
global variable at the deployment ScrubCams level rather than on the
individual ScrubCam level.